Add “Edit link” support
